+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Verb Changelog
-## master
+## **2.5.0** - 2020-01-30
 - Added `verb-trim-body-end` variable.
 - Extended `:op send` so that it now accepts an optional, additional `get-headers` or `get-body` argument.
 - Added "Customize Verb" button.